Extend basic pipeline spec, incl. tasks/execs, jobs, artifact, fetch and docs
Added docs what the basic pipeline should do and what each item does.

Introduced jobs and execs (which are equal to tasks in most systems),
this to be able to run stuff in parallel within a stage, but also
still be able to execute multiple commands inside a job in serial.
This comes basicly from go.cd, but is also mentioned in the CD book.

And added a way to publish and fetch artifacts, so they can be
reused between stages